requirements.txt
keras==3.0.5
matplotlib==3.8.2
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
scikit-learn==1.4.0
tensorflow==2.15.0
tf-keras==2.15.0
